Moving components + datapanel height fix
