+++ src/components/documentBuilder/DocumentContext.tsx
@@ -1,17 +1,18 @@
 import React from "react";
-import { BlockOptions } from "@/core";
+import { BlockArgContext, BlockOptions } from "@/core";
 import ConsoleLogger from "@/tests/ConsoleLogger";
 
-type InnerContext = {
-  options: BlockOptions;
+type DocumentState = {
+  options: BlockOptions<BlockArgContext>;
 };
 
-const initialValue: InnerContext = {
+const initialValue: DocumentState = {
   options: {
+    // eslint-disable-next-line @typescript-eslint/consistent-type-assertions -- have to cast to BlockArgContext explicitly
     ctxt: {
       "@input": {},
       "@options": {},
-    },
+    } as BlockArgContext,
     // The root should correspond to the host page's content script. If we passed document here, it would end up being
     // the document what's rendering the document (e.g., the sidebar panel's iframe document)
     root: null,
@@ -20,6 +21,6 @@ const initialValue: InnerContext = {
   },
 };
 
-const InnerComponentContext = React.createContext<InnerContext>(initialValue);
+const DocumentContext = React.createContext<DocumentState>(initialValue);
 
-export default InnerComponentContext;
+export default DocumentContext;
