+++ src/core.ts
@@ -225,11 +225,11 @@ export interface Logger {
 
 export type ReaderRoot = HTMLElement | Document;
 
-export type BlockOptions = {
-  // Using "any" for now so that blocks don't have to assert/cast all their argument types. We're checking
-  // the inputs using yup/jsonschema, so the types should match what's expected.
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  ctxt: Record<string, any>;
+// Using "any" for now so that blocks don't have to assert/cast all their argument types. We're checking
+// the inputs using yup/jsonschema, so the types should match what's expected.
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export type BlockOptions<TCtxt = Record<string, any>> = {
+  ctxt: TCtxt;
   logger: Logger;
   root: ReaderRoot;
   headless?: boolean;
