+++ src/devTools/editor/tabs/editTab/dataPanel/DataPanel.module.scss
@@ -19,6 +19,10 @@
   display: flex;
   flex-direction: column;
   height: 100%;
+
+  :global(.tab-pane) {
+    min-height: 100%;
+  }
 }
 
 .tabNav {
